Handler and query helpers in controllers package
