+++ contracts/cw4-group/src/contract.rs
@@ -5,7 +5,7 @@ use cosmwasm_std::{
 use cw0::maybe_canonical;
 use cw2::set_contract_version;
 use cw4::{
-    HooksResponse, Member, MemberChangedHookMsg, MemberDiff, MemberListResponse, MemberResponse,
+    Member, MemberChangedHookMsg, MemberDiff, MemberListResponse, MemberResponse,
     TotalWeightResponse,
 };
 use cw_storage_plus::Bound;
@@ -64,8 +64,8 @@ pub fn handle(
         HandleMsg::UpdateMembers { add, remove } => {
             handle_update_members(deps, env, info, add, remove)
         }
-        HandleMsg::AddHook { addr } => handle_add_hook(deps, info, addr),
-        HandleMsg::RemoveHook { addr } => handle_remove_hook(deps, info, addr),
+        HandleMsg::AddHook { addr } => Ok(HOOKS.handle_add_hook(&ADMIN, deps, info, addr)?),
+        HandleMsg::RemoveHook { addr } => Ok(HOOKS.handle_remove_hook(&ADMIN, deps, info, addr)?),
     }
 }
 
@@ -126,26 +126,6 @@ pub fn update_members(
     Ok(MemberChangedHookMsg { diffs })
 }
 
-pub fn handle_add_hook(
-    deps: DepsMut,
-    info: MessageInfo,
-    addr: HumanAddr,
-) -> Result<HandleResponse, ContractError> {
-    ADMIN.assert_admin(deps.as_ref(), &info.sender)?;
-    HOOKS.add_hook(deps.storage, addr)?;
-    Ok(HandleResponse::default())
-}
-
-pub fn handle_remove_hook(
-    deps: DepsMut,
-    info: MessageInfo,
-    addr: HumanAddr,
-) -> Result<HandleResponse, ContractError> {
-    ADMIN.assert_admin(deps.as_ref(), &info.sender)?;
-    HOOKS.remove_hook(deps.storage, addr)?;
-    Ok(HandleResponse::default())
-}
-
 pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
     match msg {
         QueryMsg::Member {
@@ -155,15 +135,10 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
         QueryMsg::ListMembers { start_after, limit } => {
             to_binary(&list_members(deps, start_after, limit)?)
         }
-        QueryMsg::Admin {} => to_binary(&ADMIN.query_admin(deps)?),
         QueryMsg::TotalWeight {} => to_binary(&query_total_weight(deps)?),
-        QueryMsg::Hooks {} => to_binary(&query_hooks(deps)?),
+        QueryMsg::Admin {} => to_binary(&ADMIN.query_admin(deps)?),
+        QueryMsg::Hooks {} => to_binary(&HOOKS.query_hooks(deps)?),
     }
-}
-
-fn query_hooks(deps: Deps) -> StdResult<HooksResponse> {
-    let hooks = HOOKS.may_load(deps.storage)?.unwrap_or_default();
-    Ok(HooksResponse { hooks })
 }
 
 fn query_total_weight(deps: Deps) -> StdResult<TotalWeightResponse> {
@@ -395,7 +370,7 @@ mod tests {
         let mut deps = mock_dependencies(&[]);
         do_init(deps.as_mut());
 
-        let hooks = query_hooks(deps.as_ref()).unwrap();
+        let hooks = HOOKS.query_hooks(deps.as_ref()).unwrap();
         assert!(hooks.hooks.is_empty());
 
         let contract1 = HumanAddr::from("hook1");
@@ -414,7 +389,7 @@ mod tests {
             add_msg.clone(),
         )
         .unwrap_err();
-        assert_eq!(err, AdminError::NotAdmin {}.into());
+        assert_eq!(err, HookError::Admin(AdminError::NotAdmin {}).into());
 
         // admin can add it, and it appears in the query
         let admin_info = mock_info(INIT_ADMIN, &[]);
@@ -425,7 +400,7 @@ mod tests {
             add_msg.clone(),
         )
         .unwrap();
-        let hooks = query_hooks(deps.as_ref()).unwrap();
+        let hooks = HOOKS.query_hooks(deps.as_ref()).unwrap();
         assert_eq!(hooks.hooks, vec![contract1.clone()]);
 
         // cannot remove a non-registered contract
@@ -446,7 +421,7 @@ mod tests {
             addr: contract2.clone(),
         };
         let _ = handle(deps.as_mut(), mock_env(), admin_info.clone(), add_msg2).unwrap();
-        let hooks = query_hooks(deps.as_ref()).unwrap();
+        let hooks = HOOKS.query_hooks(deps.as_ref()).unwrap();
         assert_eq!(hooks.hooks, vec![contract1.clone(), contract2.clone()]);
 
         // cannot re-add an existing contract
@@ -470,7 +445,7 @@ mod tests {
             remove_msg.clone(),
         )
         .unwrap_err();
-        assert_eq!(err, AdminError::NotAdmin {}.into());
+        assert_eq!(err, HookError::Admin(AdminError::NotAdmin {}).into());
 
         // remove the original
         let _ = handle(
@@ -480,7 +455,7 @@ mod tests {
             remove_msg.clone(),
         )
         .unwrap();
-        let hooks = query_hooks(deps.as_ref()).unwrap();
+        let hooks = HOOKS.query_hooks(deps.as_ref()).unwrap();
         assert_eq!(hooks.hooks, vec![contract2.clone()]);
     }
 
@@ -489,7 +464,7 @@ mod tests {
         let mut deps = mock_dependencies(&[]);
         do_init(deps.as_mut());
 
-        let hooks = query_hooks(deps.as_ref()).unwrap();
+        let hooks = HOOKS.query_hooks(deps.as_ref()).unwrap();
         assert!(hooks.hooks.is_empty());
 
         let contract1 = HumanAddr::from("hook1");
